backend/requirements.txt
# FastAPI framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# CORS and utilities
python-multipart>=0.0.6
email-validator>=2.1.0
